+++ src/vscp/daemon/controlobject.cpp
@@ -31,8 +31,8 @@
 #define _POSIX
 
 #ifdef WIN32
-#include <nb30.h>
 #include <stdAfx.h>
+#include <nb30.h>
 #endif
 
 #include <controlobject.h>
@@ -851,8 +851,8 @@ clock_gettime(int X, struct timeval *tv)
 
   t.QuadPart -= offset.QuadPart;
   microseconds = (double) t.QuadPart / frequencyToMicroseconds;
-  t.QuadPart   = microseconds;
-  tv->tv_sec   = t.QuadPart / 1000000;
+  t.QuadPart   = (LONGLONG)microseconds;
+  tv->tv_sec   = (long)t.QuadPart / 1000000;
   tv->tv_usec  = t.QuadPart % 1000000;
   return (0);
 }
